:construction: account restore selection removed. unused code deleted
