--- Classes/AlgorandSecureBackup/AccountList/Views/AccountsHeader/AlgorandSecureBackupAccountListAccountsHeader.swift
@@ -1,50 +1,0 @@
-// Copyright 2023 Pera Wallet, LDA
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-
-//    http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//   AlgorandSecureBackupAccountListAccountsHeader.swift
-
-import Foundation
-import MacaroonUIKit
-import UIKit
-
-final class AlgorandSecureBackupAccountListAccountsHeader:
-    CollectionCell<AlgorandSecureBackupAccountListAccountsHeaderView>,
-    ViewModelBindable,
-    UIInteractable {
-    var state: AlgorandSecureBackupAccountListAccountHeaderItemState = .selectAll {
-        didSet { updateStateIfNeeded(old: oldValue) }
-    }
-
-    override class var contextPaddings: LayoutPaddings {
-        return (0, 24, 20, 24)
-    }
-
-    static let theme = AlgorandSecureBackupAccountListAccountsHeaderViewTheme()
-
-    override init(
-        frame: CGRect
-    ) {
-        super.init(frame: frame)
-
-        contextView.customize(Self.theme)
-    }
-}
-
-extension AlgorandSecureBackupAccountListAccountsHeader {
-    private func updateStateIfNeeded(old: AlgorandSecureBackupAccountListAccountHeaderItemState) {
-        if state != old {
-            contextView.updateState(state)
-        }
-    }
-}
