+++ Classes/Routing/Screen.swift
@@ -367,7 +367,7 @@ indirect enum Screen {
         selectedAccounts: [Account],
         eventHandler: WebImportSuccessScreen.EventHandler
     )
-    case algorandSecureBackupRestoreAccountList(accountImportParameters: [AccountImportParameters], eventHandler: AlgorandSecureBackupAccountListScreen.EventHandler)
+//    case algorandSecureBackupRestoreAccountList(accountImportParameters: [AccountImportParameters], eventHandler: AlgorandSecureBackupAccountListScreen.EventHandler)
     case algorandSecureBackupRecoverMnemonic(backup: SecureBackup, eventHandler: AlgorandSecureBackupRecoverMnemonicScreen.EventHandler)
     case importTextDocumentPicker(delegate: UIDocumentPickerDelegate)
     case buySellOptions(eventHandler: BuySellOptionsScreen.EventHandler)
